Adjust keys to left/right arrows for skip/keep
Since the text boxes for the skip/keep buttons are bound to ax1, we cannot destroy this axis (as I had done previously). This change keeps each of the axes, and finds them if the figure already exists. If the figure is being created for the first time, it'll make the axes in the correct orientation. Also added the feature to press escape and raise an exception, which will exit the checking process.