Changed .kml into .geojson
We have decided to move away from using .KML files as they do not contain the spatial reference and some users had troubles using them in GIS applications. Instead the files are now saved in .geojson which is a great format as it is human-readable and includes the spatial reference.

All functions and documentation have been updated to use the .geojson format (with geopandas).