joblib and sklearn upgrade
I have updated the environment.yml file to use the latest skicit-learn package as there were issues with a deprecated version of joblib.
Now the issue that persists is that the neural network models (.pkl) under /classification/models are not compatible with the new version of joblib and cannot be loaded.


Former-commit-id: 860d9f021b88b8d0f6568246c685f687418b7428